+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 0.5.0 [2020-06-26]
+
+### Changed
+
+- `.override` static method of formatters has been renamed to `.wrap`.
+- `.wrap` now defaults delegation of `suggestionTools`.
+
 ## O.4.0 [2020-06-25]
 
 ### Added